middleware/requirements.txt
flask
flask-cors
openai
replicate
waitress
Flask-Compress
boto3
python-dotenv